pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spam-analytics"
version = "0.1.0"
description = "Spam email analytics toolkit with preprocessing, modeling, and visualization utilities."
authors = [
  { name = "HW3 Team" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pandas>=2.1",
  "numpy>=1.25",
  "scikit-learn>=1.3",
  "matplotlib>=3.7",
  "seaborn>=0.13",
  "plotly>=5.18",
  "typer>=0.12",
  "rich>=13.7",
  "streamlit>=1.31",
  "joblib>=1.3",
  "pyyaml>=6.0",
  "nltk>=3.8"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "black>=23.9",
  "isort>=5.12"
]

[project.scripts]
spam-analytics = "spam_analytics.cli:app"
